% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_covid_hub_forecasts.R
\name{as_covid_hub_forecasts}
\alias{as_covid_hub_forecasts}
\title{Reformat model outputs stored as a \code{model_output_tbl} class to that of a
\code{data.frame} formatted according to standards of the COVID-19 Forecasting
Hub which can be processed by functions from the \code{covidHubUtils} package
such as \code{score_forecasts()} or \code{plot_forecasts()}. The supplied
\code{model_output_tbl} should have columns defining properties akin to
reference dates, locations, horizons, and targets.}
\usage{
as_covid_hub_forecasts(model_outputs)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_tbl} with component
model outputs (e.g., predictions). Should have columns containing the
following information: model name, reference date, location, horizon,
target, output type, output type id, and value. The target column should
contain targets of the form "\link{temporal resolution} \link{target}" or
"\link{temporal resolution} ahead \link{target}", such as "wk ahead inc flu hosp"
"wk inc flu hosp", to properly create a temporal resolution column.}
}
\value{
a \code{data.frame} of reformatted model outputs that may be fed into
any of the \code{covidHubUtils} functions with 10 total columns: model,
forecast_date, location, horizon, temporal_resolution, target_variable,
target_end_date, type, quantile, value.
}
\description{
Reformat model outputs stored as a \code{model_output_tbl} class to that of a
\code{data.frame} formatted according to standards of the COVID-19 Forecasting
Hub which can be processed by functions from the \code{covidHubUtils} package
such as \code{score_forecasts()} or \code{plot_forecasts()}. The supplied
\code{model_output_tbl} should have columns defining properties akin to
reference dates, locations, horizons, and targets.
}
