% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{evaluate_flu_scores}
\alias{evaluate_flu_scores}
\title{Evaluate forecasts (from scores) across different evaluation groupings}
\usage{
evaluate_flu_scores(scores, grouping_variables, baseline_name, us_only = FALSE)
}
\arguments{
\item{scores}{A data frame of forecast scores to be summarized. Must have
columns that match the output from \code{covidHubUtils::score_forecasts()} and
include scores for a baseline model.}

\item{grouping_variables}{A character vector specifying the type of grouping
to be used in evaluation. Options are \code{season}, \code{horizon}, \code{forecast_week},
\code{location}.}

\item{baseline_name}{A character string specifying the name of the baseline
model to calculate relative metrics against}

\item{us_only}{A boolean specifying whether to summarize metrics for the US
national level only or just states}
}
\value{
A data frame of summarized forecast score metrics across all
provided forecasts
}
\description{
Evaluate forecasts (from scores) across different evaluation groupings
}
