% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cells.R
\name{format_cells}
\alias{format_cells}
\title{Apply formatting to cells in a data frame}
\usage{
format_cells(
  df,
  rows,
  cols,
  value = c("italics", "bold", "strikethrough", "code")
)
}
\arguments{
\item{df}{a data.frame}

\item{rows}{rows to format}

\item{cols}{columns to format}

\item{value}{type of formatting to apply}
}
\description{
Function adapted from answered question on Stack Overflow.
Source: https://stackoverflow.com/a/49656650
}
