% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_truth_forecasts_functions.R
\name{generate_flu_ensemble}
\alias{generate_flu_ensemble}
\title{Compute ensemble model outputs for incident for your hospitalizations
as a quantile average, quantile median, linear pool with normal tails,
or linear pool with log normal tails for each combination of model task,
output type, and output type id. The outputs are formatted to be fed into
\code{covidHubUtils::score_forecasts()} for scoring}
\usage{
generate_flu_ensemble(
  model_outputs,
  include_baseline = FALSE,
  ensemble_type,
  tail_dist = NULL,
  ...
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_df} with component
model outputs (e.g., predictions).}

\item{include_baseline}{\code{logical} that specifies whether to include the
Flusight-baseline model in the ensemble.  Defaults to FALSE.}

\item{ensemble_type}{\code{character} string that specifies the ensembling method
to be used on the flu forecasts. Can be "mean", "median", or "linear_pool".}

\item{tail_dist}{\code{character} string that specifies the type of distribution
to use when calculating the tails for a linear pool ensemble.
Defaults to NULL. This argument is ignored for Vincentization.}

\item{...}{parameters that are passed to \code{distfromq::make_q_fun}, specifying
details of how to estimate a quantile function from provided quantile levels
and quantile values.}
}
\value{
a \code{model_out_tbl} object of ensemble predictions for flu hospitalizations.
}
\description{
Compute ensemble model outputs for incident for your hospitalizations
as a quantile average, quantile median, linear pool with normal tails,
or linear pool with log normal tails for each combination of model task,
output type, and output type id. The outputs are formatted to be fed into
\code{covidHubUtils::score_forecasts()} for scoring
}
