% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_truth_forecasts_functions.R
\name{get_flu_forecasts_single_date}
\alias{get_flu_forecasts_single_date}
\title{Pulls forecasts for incident flu hospitalizations from zoltar and
turns the data-frame into a \code{model_out_tbl} object as defined by the
hubverse to be fed into ensembling functions from \code{hubEnsembles}}
\usage{
get_flu_forecasts_single_date(zoltar_connection, project_url, origin_date)
}
\arguments{
\item{zoltar_connection}{a connection through which to access zoltar}

\item{project_url}{\code{character} string of the URL for these zoltar project
from which to query the component forecasts}

\item{origin_date}{\code{character} string that specifies the date on which
the component forecasts were made that should be used for the ensemble}
}
\value{
a \code{data.frame} of forecast data for flu hospitalizations with columns:
model_id, forecast_date, location, horizon, temporal_resolution,
target_long, output_type, output_type_id, value
}
\description{
Pulls forecasts for incident flu hospitalizations from zoltar and
turns the data-frame into a \code{model_out_tbl} object as defined by the
hubverse to be fed into ensembling functions from \code{hubEnsembles}
}
