% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_truth_forecasts_functions.R
\name{get_flu_truth_single_date}
\alias{get_flu_truth_single_date}
\title{Pulls truth data for incident flu hospitalizations from zoltar and
formats the data to be fed into \code{covidHubUtils::score_forecasts()}}
\usage{
get_flu_truth_single_date(zoltar_connection, project_url, origin_date)
}
\arguments{
\item{zoltar_connection}{a connection through which to access zoltar}

\item{project_url}{\code{character} string of the URL for these zoltar project
from which to query the component forecasts}

\item{origin_date}{\code{character} string that specifies the date on which
the component forecasts were made that should be used for the ensemble}
}
\value{
a \code{data.frame} of truth data for flu hospitalizations with columns:
model, target_variable, target_end_date, location, value
}
\description{
Pulls truth data for incident flu hospitalizations from zoltar and
formats the data to be fed into \code{covidHubUtils::score_forecasts()}
}
