% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubEnsemblesManuscript-package.R
\docType{package}
\name{hubEnsemblesManuscript-package}
\alias{hubEnsemblesManuscript}
\alias{hubEnsemblesManuscript-package}
\title{hubEnsemblesManuscript: Repository accompanying paper on hubEnsembles}
\description{
Repository containing data and analysis for a paper on hubEnsembles, a hubverse package of ensemble methods for combining hub model outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Infectious-Disease-Modeling-Hubs/hubEnsemblesManuscript}
}

}
\author{
\strong{Maintainer}: Li Shandross \email{lshandross@umass.edu} (\href{https://orcid.org/0009-0008-1348-1954}{ORCID})

Authors:
\itemize{
  \item Emily Howerton \email{ehowerton@psu.edu} (\href{https://orcid.org/0000-0002-0639-3728}{ORCID})
  \item Evan L Ray \email{elray@umass.edu}
}

Other contributors:
\itemize{
  \item Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID}) [contributor]
}

}
\keyword{internal}
