% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_evaluated_scores_forecast_date}
\alias{plot_evaluated_scores_forecast_date}
\title{Plot summarized metrics against forecast_date, with option to plot averaged
target data on the same graph}
\usage{
plot_evaluated_scores_forecast_date(
  summarized_scores,
  model_names,
  model_colors,
  y_var = "wis",
  h = 1,
  main = NULL,
  time_series_data = NULL,
  time_series_scaling = 0.125
)
}
\arguments{
\item{summarized_scores}{A data frame of summarized scores. Must contain one
row per model and horizon week combination plus a \code{horizon} column}

\item{model_names}{An ordered vector of model names}

\item{model_colors}{An ordered vector of model colors. Must match with
\code{model_names} order}

\item{y_var}{A string specifying which metric to plot as the y-variable}

\item{h}{An integer specifying the horizon of the desired metric to plot}

\item{main}{A string specifying the plot title. Defaults to NULL.}

\item{time_series_data}{A data frame of time series target data to plot in the
same figure. Defaults to \code{NULL}. If provided must contain target_end_date and
value columns.}

\item{time_series_scaling}{A numeric specifying a multiplier for the time series
values so that it fits nicely on the graph.}
}
\value{
A scatter plot (with observations connected by lines) of the
specified summary metric vs forecast date
}
\description{
Plot summarized metrics against forecast_date, with option to plot averaged
target data on the same graph
}
