% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_flu_time_series}
\alias{plot_flu_time_series}
\title{Plot averaged target data against forecast dates}
\usage{
plot_flu_time_series(
  time_series_data,
  date_range = NULL,
  main = "target data",
  plot_color = "black"
)
}
\arguments{
\item{time_series_data}{A data frame of target time series data. Must contain a
target_end_date column}

\item{date_range}{An ordered string vector giving the range of dates to plot}

\item{main}{A string specifying the plot title. Defaults to NULL.}

\item{plot_color}{A string specifying the color that the target data should be
plotted as. Defaults to "black".}
}
\value{
A scatter plot (with observations connected by lines) of the target
data vs forecast date
}
\description{
Plot averaged target data against forecast dates
}
