% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_flu_truth}
\alias{plot_flu_truth}
\title{Plot summarized metrics against forecast date}
\usage{
plot_flu_truth(
  truth_data,
  date_range = NULL,
  main = "truth data",
  plot_color = "black"
)
}
\arguments{
\item{truth_data}{A data frame of truth data. Must contain a target_end_date
column}

\item{date_range}{An ordered string vector giving the range of dates to plot}

\item{main}{A string specifying the plot title. Defaults to NULL.}

\item{plot_color}{A string specifying the color that the truth should be
plotted as. Defaults to "black".}
}
\value{
A scatter plot (with observations connected by lines) of the truth
data vs forecast date
}
\description{
Plot summarized metrics against forecast date
}
