% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_truth_forecasts_functions.R
\name{wrapper_flu_ensemble_single_date}
\alias{wrapper_flu_ensemble_single_date}
\title{Wrapper function used to compute ensemble model outputs as a
quantile average, quantile median, linear pool with normal tails,
or linear pool with log normal tails for each combination model task,
output type, and output type id. Component model forecasts for
incident flu hospitalizations are queried from zoltar with ensemble
outputs formatted to be fed into \code{covidHubUtils::score_forecasts()}
for scoring}
\usage{
wrapper_flu_ensemble_single_date(
  zoltar_connection,
  project_url,
  origin_date,
  include_baseline = FALSE,
  ensemble_type,
  tail_dist = NULL,
  ...
)
}
\arguments{
\item{zoltar_connection}{a connection through which to access zoltar}

\item{project_url}{\code{character} string of the URL for these zoltar project
from which to query the component forecasts}

\item{origin_date}{\code{character} string that specifies the date on which
the component forecasts were made that should be used for the ensemble}

\item{include_baseline}{\code{logical} that specifies whether to include the
Flusight-baseline model in the ensemble.  Defaults to FALSE.}

\item{ensemble_type}{\code{character} string that specifies the ensembling method
to be used on the flu forecasts. Can be "mean", "median", or "linear_pool".}

\item{tail_dist}{\code{character} string that specifies the type of distribution
to use when calculating the tails for a linear pool ensemble.
Defaults to NULL. This argument is ignored for Vincentization.}

\item{...}{parameters that are passed to \code{distfromq::make_q_fun}, specifying
details of how to estimate a quantile function from provided quantile levels
and quantile values.}
}
\value{
a \code{model_out_tbl} object of ensemble predictions for flu hospitalizations.
}
\description{
Wrapper function used to compute ensemble model outputs as a
quantile average, quantile median, linear pool with normal tails,
or linear pool with log normal tails for each combination model task,
output type, and output type id. Component model forecasts for
incident flu hospitalizations are queried from zoltar with ensemble
outputs formatted to be fed into \code{covidHubUtils::score_forecasts()}
for scoring
}
