% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoltar_to_model_out.R
\name{zoltar_to_model_out}
\alias{zoltar_to_model_out}
\title{Reformat a data frame of forecasts stored in the zoltar format to that of a
\code{model_output_tbl} object defined by the hubverse}
\usage{
zoltar_to_model_out(zoltar_forecasts)
}
\arguments{
\item{model_outputs}{a data frame of forecasts formatted according to zoltar
standards with the following columns: model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value.}
}
\value{
a \code{model_out_tbl} of reformatted forecasts with the following columns:
model_id, forecast_date, location, horizon, target, target_end_date,
output_type, output_type_id, value.
}
\description{
Reformat a data frame of forecasts stored in the zoltar format to that of a
\code{model_output_tbl} object defined by the hubverse
}
